# lib/api-zod/src/generated/types/gameAction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type GameAction = {
  type: 'hit' | 'stand' | 'double' | 'split' | 'surrender';
  /** @minimum 0 */
  seatId: number;
  handId: string;
} | {
  type: 'insurance' | 'decline_insurance';
  /** @minimum 0 */
  seatId: number;
};