# lib/api-zod/src/generated/types/gameResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameResponseEventsItem } from './gameResponseEventsItem';
import type { GameResponseGame } from './gameResponseGame';

export interface GameResponse {
  sessionId: string;
  tableId: string;
  /** @minimum 0 */
  round: number;
  /** @minimum 0 */
  version: number;
  /** @minimum 0 */
  balance: number;
  /** Public game state; the private shoe and unrevealed hole card are never returned. */
  game: GameResponseGame;
  /** Ordered card and phase events for client animation. */
  events: GameResponseEventsItem[];
}