# lib/api-zod/src/generated/types/seatBet.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeatBetSideBets } from './seatBetSideBets';

export interface SeatBet {
  /** @minimum 0 */
  seatId: number;
  /** @minimum 1 */
  main: number;
  sideBets?: SeatBetSideBets;
}